extract feature set primitives into dedicated core crate
