+++ Cargo.toml
@@ -10,6 +10,7 @@ members = [
   "crates/bitnet-runtime-context-core",
   "crates/bitnet-bdd-grid",
   "crates/bitnet-bdd-grid-core",
+  "crates/bitnet-feature-set-core", # SRP: feature taxonomy and feature-set operations
   "crates/bitnet-compat",
   "crates/bitnet-ggml-ffi",     # GGML FFI for IQ2_S support
   "crates/bitnet-runtime-feature-flags",
@@ -153,6 +154,7 @@ default-members = [
   # Testing infrastructure crates (always built/tested by default)
   "crates/bitnet-bdd-grid",
   "crates/bitnet-bdd-grid-core",
+  "crates/bitnet-feature-set-core", # SRP: feature taxonomy and feature-set operations
   "crates/bitnet-feature-contract",
   "crates/bitnet-feature-matrix",
   "crates/bitnet-server-health-types-core",
