+++ crates/bitnet-feature-set-core/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "bitnet-feature-set-core"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+keywords.workspace = true
+categories.workspace = true
+description = "Core BitNet feature taxonomy and feature-set primitives"
+include = ["src/**", "Cargo.toml"]
+
+[dependencies]
